+++ CHANGELOG.md
@@ -10,8 +10,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Changes
 
-- Input: added onChange prop to explicitly allow it to be used as a controlled component
-- Select: selectedOption holds the selected value, instead of just the default.
+- `Select`: selectedOption holds the selected value, instead of just the default.
+- `Hero`: `Hero.Secondary` now changes background color based on NYPL site section
+- `Hero`: All props are now optional
+- `Input`: added onChange prop to explicitly allow it to be used as a controlled component
 - Uses stricter linters and more Prettier configs.
 
 ### Breaking Changes